Renamed DebuggingInformation into BuildDebuggingInformation
